requirements.txt
Flask
transformers
torch
faiss-cpu
numpy
pandas
